Fix frame pacing of interpolation by fixing the interval to 35hz
Previously I was doing the timing based on the start of the gametic.
Unfortunately, in certain frame paces (Especially 70hz vsync) this
causes a stutter because we will run 2 interpolated frames inside a
gametic but sometimes we would only run one, and reset the gametic each
time. This would cause a frame to be rerendered with the same
interpolation every few frames.

This approach instead calculated the interpolation from a fixed base
interval which never changes, so no matter when we do a gametic we will
have the next interpolation properly paced.