Rework the input buffer with interpolation on
Previously when interpolation was on there were some buffer
synchronization issues. This was somewhat resolved with polling
the input buffer on every game tick directly, but this
doesn't handle rapid events like mouse movement properly,
this adjusts the code to use the same mechanism as the
capped mode. One minor difference is that we synchronize
maketic and gametic if we are stalling for updates under
the assumption that the gametics are caught up